requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wsproto==1.2.0
XlsxWriter==3.2.0
